+++ src/funding/common.jsx
@@ -4,7 +4,7 @@
 
 import { node } from 'jsx-pragmatic';
 import type { FundingEligibilityType } from '@paypal/sdk-client/src';
-import { PLATFORM, type LocaleType, COUNTRY, CARD, COMPONENTS } from '@paypal/sdk-constants/src';
+import { PLATFORM, type LocaleType, COUNTRY, CARD, COMPONENTS, FUNDING } from '@paypal/sdk-constants/src';
 import { type ChildType } from 'jsx-pragmatic/src';
 import { LOGO_COLOR } from '@paypal/sdk-logos/src';
 
@@ -29,6 +29,7 @@ export type LogoOptions = {|
 |};
 
 export type LabelOptions = {|
+    i : number,
     logo : ChildType,
     label : ?$Values<typeof BUTTON_LABEL>,
     locale : LocaleType,
@@ -64,24 +65,29 @@ export type FundingSourceConfig = {|
     maxCards? : { [$Values<typeof COUNTRY>] : number },
     remembered? : boolean,
     vendors? : { [$Values<typeof CARD>] : ?CardConfig },
-    eligible? : ({ components : $ReadOnlyArray<$Values<typeof COMPONENTS>>, fundingEligibility : FundingEligibilityType, layout : ?$Values<typeof BUTTON_LAYOUT> }) => boolean,
+    eligible? : ({ components : $ReadOnlyArray<$Values<typeof COMPONENTS>>, fundingEligibility : FundingEligibilityType, fundingSource : ?$Values<typeof FUNDING>, layout : ?$Values<typeof BUTTON_LAYOUT> }) => boolean,
     Logo : (LogoOptions) => ChildType,
     Label : (LabelOptions) => ChildType,
     VaultLabel? : (VaultLabelOptions) => ChildType,
     Tag? : (TagOptions) => ?ChildType,
     colors : $ReadOnlyArray<$Values<typeof BUTTON_COLOR>>,
+    textColors : { [$Values<typeof BUTTON_COLOR>] : $Values<typeof BUTTON_COLOR> },
     secondaryColors : { [$Values<typeof BUTTON_COLOR>] : $Values<typeof BUTTON_COLOR> },
     secondaryVaultColors : { [$Values<typeof BUTTON_COLOR>] : $Values<typeof BUTTON_COLOR> },
     logoColors : { [$Values<typeof BUTTON_COLOR>] : $Values<typeof LOGO_COLOR> },
     shapes : $ReadOnlyArray<$Values<typeof BUTTON_SHAPE>>
 |};
 
-export function BasicLabel({ logo, label, layout, multiple, period, locale: { lang } } : LabelOptions) : ChildType {
+export function BasicLabel({ i, logo, label, layout, multiple, period, locale: { lang } } : LabelOptions) : ChildType {
     if (layout === BUTTON_LAYOUT.HORIZONTAL && multiple) {
         return logo;
     }
 
     if (__WEB__) {
+        return logo;
+    }
+
+    if (i > 0) {
         return logo;
     }
 
@@ -135,6 +141,13 @@ export const DEFAULT_FUNDING_CONFIG : FundingSourceConfig = {
         BUTTON_SHAPE.RECT,
         BUTTON_SHAPE.PILL
     ],
+
+    textColors: {
+        [DEFAULT]:               BUTTON_COLOR.BLACK,
+        [BUTTON_COLOR.BLUE]:     BUTTON_COLOR.WHITE,
+        [BUTTON_COLOR.BLACK]:    BUTTON_COLOR.WHITE,
+        [BUTTON_COLOR.DARKBLUE]: BUTTON_COLOR.WHITE
+    },
 
     secondaryColors: {
         [ DEFAULT ]:            BUTTON_COLOR.SILVER,