+++ src/zoid/wallet/component.jsx
@@ -2,18 +2,24 @@
 /** @jsx node */
 /* eslint max-lines: 0 */
 
-import { getLogger, getPayPalDomainRegex, getSDKMeta, getPayPalDomain, getClientID } from '@paypal/sdk-client/src';
+import { node, dom } from 'jsx-pragmatic/src';
+import { getLogger, getPayPalDomainRegex, getSDKMeta, getPayPalDomain, getClientID, getUserAccessToken, getClientAccessToken, getUserAuthCode, getLocale, getPartnerAttributionID, getCorrelationID, getSessionID, getEnv, getStageHost, getAPIStageHost, getPlatform, getCurrency, getIntent, getBuyerCountry, getCommit, getVault, getMerchantID, getCSPNonce, getDebug } from '@paypal/sdk-client/src';
 import { create, type ZoidComponent } from 'zoid/src';
-import { inlineMemoize, memoize } from 'belter/src';
+import { inlineMemoize, memoize, uniqueID } from 'belter/src';
 import { ZalgoPromise } from 'zalgo-promise/src';
+import { FUNDING } from '@paypal/sdk-constants/src';
+import { getRefinedFundingEligibility, rememberFunding } from '@paypal/funding-components/src';
+import { collectRiskData } from '@paypal/risk-data-collector/src';
 
 import { type WalletProps } from './props';
+import { WalletPrerender } from './prerender';
+import { WalletContainer } from './container';
 
 export function getWalletComponent() : ZoidComponent<WalletProps> {
     return inlineMemoize(getWalletComponent, () => {
         return create({
             tag:    'paypal-wallet',
-            url:    () => `${ getPayPalDomain() }${ window.__CHECKOUT_URI__ || __PAYPAL_CHECKOUT__.__URI__.__WALLET__ }`,
+            url:    () => `${ getPayPalDomain() }${ __PAYPAL_CHECKOUT__.__URI__.__WALLET__ }`,
             domain: getPayPalDomainRegex(),
             
             autoResize: {
@@ -28,8 +34,16 @@ export function getWalletComponent() : ZoidComponent<WalletProps> {
 
             logger: getLogger(),
 
-            prerenderTemplate: () => {
-                return null;
+            containerTemplate: ({ props, doc, uid, frame, prerenderFrame, event }) => {
+                return (
+                    <WalletContainer uid={ uid } frame={ frame } prerenderFrame={ prerenderFrame } event={ event } props={ props } />
+                ).render(dom({ doc }));
+            },
+
+            prerenderTemplate: ({ props, doc }) => {
+                return (
+                    <WalletPrerender nonce={ props.nonce } />
+                ).render(dom({ doc }));
             },
 
             attributes: {
@@ -51,6 +65,33 @@ export function getWalletComponent() : ZoidComponent<WalletProps> {
                     queryParam: true,
                     value:      getClientID
                 },
+                
+                clientAccessToken: {
+                    type:       'string',
+                    required:   false,
+                    queryParam: true,
+                    value:      getClientAccessToken
+                },
+
+                buyerAccessToken: {
+                    type:       'string',
+                    queryParam: true,
+                    required:   false,
+                    value:      getUserAccessToken
+                },
+
+                buyerAuthCode: {
+                    type:       'string',
+                    queryParam: true,
+                    required:   false,
+                    value:      getUserAuthCode
+                },
+
+                fundingSource: {
+                    type:       'string',
+                    queryParam: true,
+                    default:    () => FUNDING.PAYPAL
+                },
 
                 style: {
                     type:     'object',
@@ -71,6 +112,142 @@ export function getWalletComponent() : ZoidComponent<WalletProps> {
 
                 onApprove: {
                     type: 'function'
+                },
+
+                riskData: {
+                    type:          'object',
+                    value:         collectRiskData,
+                    queryParam:    true,
+                    serialization: 'base64'
+                },
+
+                locale: {
+                    type:       'object',
+                    queryParam: true,
+                    value:      getLocale
+                },
+
+                partnerAttributionID: {
+                    type:       'string',
+                    required:   false,
+                    value:     getPartnerAttributionID
+                },
+
+                correlationID: {
+                    type:       'string',
+                    required:   false,
+                    value:      getCorrelationID
+                },
+
+                sessionID: {
+                    type:       'string',
+                    value:      getSessionID,
+                    queryParam: true
+                },
+
+                walletSessionID: {
+                    type:       'string',
+                    value:      uniqueID,
+                    queryParam: true
+                },
+
+                env: {
+                    type:       'string',
+                    queryParam: true,
+                    value:      getEnv
+                },
+
+                stageHost: {
+                    type:       'string',
+                    value:      getStageHost,
+                    required:   false
+                },
+
+                apiStageHost: {
+                    type:       'string',
+                    value:      getAPIStageHost,
+                    required:   false
+                },
+
+                fundingEligibility: {
+                    type:          'object',
+                    value:         getRefinedFundingEligibility,
+                    queryParam:    true,
+                    serialization: 'base64'
+                },
+
+                platform: {
+                    type:       'string',
+                    queryParam: true,
+                    value:      getPlatform
+                },
+
+                remember: {
+                    type:  'function',
+                    value: () => {
+                        return (fundingSources : $ReadOnlyArray<$Values<typeof FUNDING>>) =>
+                            rememberFunding(fundingSources, { cookie: false });
+                    }
+                },
+
+                currency: {
+                    type:       'string',
+                    queryParam: true,
+                    value:      getCurrency
+                },
+
+                intent: {
+                    type:       'string',
+                    queryParam: true,
+                    value:      getIntent
+                },
+
+                buyerCountry: {
+                    type:       'string',
+                    queryParam: true,
+                    required:   false,
+                    value:      getBuyerCountry
+                },
+
+                commit: {
+                    type:       'boolean',
+                    queryParam: true,
+                    value:      getCommit
+                },
+
+                vault: {
+                    type:       'boolean',
+                    queryParam: true,
+                    value:      getVault
+                },
+            
+                merchantID: {
+                    type:       'array',
+                    queryParam: true,
+                    value:      getMerchantID
+                },
+
+                csp: {
+                    type:     'object',
+                    required: false,
+                    value:    () => {
+                        return {
+                            nonce: getCSPNonce()
+                        };
+                    }
+                },
+            
+                getPageUrl: {
+                    type:  'function',
+                    value: () => {
+                        return () => window.location.href;
+                    }
+                },
+
+                debug: {
+                    type:       'boolean',
+                    value:      getDebug,
+                    queryParam: true
                 }
             }
         });