+++ src/zoid/wallet/props.js
@@ -1,5 +1,8 @@
 /* @flow */
 
+import { FUNDING } from '@paypal/sdk-constants/src';
+
 export type WalletProps = {|
-    
+    nonce : ?string,
+    fundingSource : $Values<typeof FUNDING>
 |};